get rid of github insights file, finish fixup
